Moved to io.smallrye.common.io.jar package
