+++ io/src/main/java/io/smallrye/common/io/jar/JarFiles.java
@@ -1,12 +1,11 @@
-package io.smallrye.common.io;
+package io.smallrye.common.io.jar;
 
 import java.io.File;
 import java.io.IOException;
 import java.util.jar.JarFile;
-import java.util.zip.ZipFile;
 
 /**
- * Java 9+ variant of a JDK-specific class for working with {@code JarFile}s.
+ * Java 8 variant of a JDK-specific class for working with {@code JarFile}s.
  */
 public class JarFiles {
     /**
@@ -14,7 +13,7 @@ public class JarFiles {
      * On Java 9+, an equivalent that is multi-release-enabled is returned.
      */
     public static JarFile create(String name) throws IOException {
-        return new JarFile(new File(name), true, ZipFile.OPEN_READ, JarFile.runtimeVersion());
+        return new JarFile(name);
     }
 
     /**
@@ -22,7 +21,7 @@ public class JarFiles {
      * On Java 9+, an equivalent that is multi-release-enabled is returned.
      */
     public static JarFile create(String name, boolean verify) throws IOException {
-        return new JarFile(new File(name), verify, ZipFile.OPEN_READ, JarFile.runtimeVersion());
+        return new JarFile(name, verify);
     }
 
     /**
@@ -30,7 +29,7 @@ public class JarFiles {
      * On Java 9+, an equivalent that is multi-release-enabled is returned.
      */
     public static JarFile create(File file) throws IOException {
-        return new JarFile(file, true, ZipFile.OPEN_READ, JarFile.runtimeVersion());
+        return new JarFile(file);
     }
 
     /**
@@ -38,6 +37,6 @@ public class JarFiles {
      * On Java 9+, an equivalent that is multi-release-enabled is returned.
      */
     public static JarFile create(File file, boolean verify) throws IOException {
-        return new JarFile(file, verify, ZipFile.OPEN_READ, JarFile.runtimeVersion());
+        return new JarFile(file, verify);
     }
 }
