+++ io/src/main/java10/io/smallrye/common/io/jar/JarEntries.java
@@ -1,4 +1,4 @@
-package io.smallrye.common.io;
+package io.smallrye.common.io.jar;
 
 import java.util.jar.JarEntry;
