+++ io/src/main/java9/io/smallrye/common/io/jar/JarFiles.java
@@ -1,11 +1,12 @@
-package io.smallrye.common.io;
+package io.smallrye.common.io.jar;
 
 import java.io.File;
 import java.io.IOException;
 import java.util.jar.JarFile;
+import java.util.zip.ZipFile;
 
 /**
- * Java 8 variant of a JDK-specific class for working with {@code JarFile}s.
+ * Java 9+ variant of a JDK-specific class for working with {@code JarFile}s.
  */
 public class JarFiles {
     /**
@@ -13,7 +14,7 @@ public class JarFiles {
      * On Java 9+, an equivalent that is multi-release-enabled is returned.
      */
     public static JarFile create(String name) throws IOException {
-        return new JarFile(name);
+        return new JarFile(new File(name), true, ZipFile.OPEN_READ, JarFile.runtimeVersion());
     }
 
     /**
@@ -21,7 +22,7 @@ public class JarFiles {
      * On Java 9+, an equivalent that is multi-release-enabled is returned.
      */
     public static JarFile create(String name, boolean verify) throws IOException {
-        return new JarFile(name, verify);
+        return new JarFile(new File(name), verify, ZipFile.OPEN_READ, JarFile.runtimeVersion());
     }
 
     /**
@@ -29,7 +30,7 @@ public class JarFiles {
      * On Java 9+, an equivalent that is multi-release-enabled is returned.
      */
     public static JarFile create(File file) throws IOException {
-        return new JarFile(file);
+        return new JarFile(file, true, ZipFile.OPEN_READ, JarFile.runtimeVersion());
     }
 
     /**
@@ -37,6 +38,6 @@ public class JarFiles {
      * On Java 9+, an equivalent that is multi-release-enabled is returned.
      */
     public static JarFile create(File file, boolean verify) throws IOException {
-        return new JarFile(file, verify);
+        return new JarFile(file, verify, ZipFile.OPEN_READ, JarFile.runtimeVersion());
     }
 }
