+++ io/src/test/java/io/smallrye/common/io/jar/JarFilesTest.java
@@ -1,4 +1,4 @@
-package io.smallrye.common.io;
+package io.smallrye.common.io.jar;
 
 import static org.junit.Assert.assertTrue;
